backend/pyproject.toml
[project]
name = "cognitive-lens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.12",
    "google>=3.0.0",
    "google-genai>=1.9.0",
    "itsdangerous>=2.2.0",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.0",
]
